requirements.txt
numpy
wheel
watchdog
django==2.2.8
django-filter
django-mailgun
djangorestframework
httpie
Markdown
mysqlclient
python-dateutil
pytz
configargparse
requests
BioPython
celery
django_celery_beat
twitter
redis
whitenoise
django_redis
tqdm
requests_futures
scipy
django_extensions
pandas
isort
psycopg2-binary
django-debug-toolbar
ete3
gunicorn
django-guardian
pylint-django
